Ajout de Phone + tests

+++ src/entities/Entreprise.java
@@ -16,7 +16,7 @@ public class Entreprise extends Contact {
 	public long getNumSiret() {
 		return numSiret;
 	}
-	public void setNumSiret(int numSiret) {
+	public void setNumSiret(long numSiret) {
 		this.numSiret = numSiret;
 	}
